Change variable labels and update documentation
Variable names have been changed for a significant number of the input,
output, and dual variables to allow for easier indexing of datasets.
The most significant changes are moving from 'l' to 'd' for
indicating loads, expanding 'pg', 'qg' to 'p_gen', 'q_gen' for the generator power
output variables, and changing line power dual variables from the format
'p_to_max' to 'pto_max'. Note that notation for similar output and dual variables
are not the same, to make it easier to parse output data and dual data columns
more easily without having to add significant length to column labels.

The default voltage variables stored in OPFLearn results are changed from
'v_bus', the complex bus voltage, to 'vm_bus' and 'va_bus', the bus voltage
magnitude and angle. This is to better align with how voltage data is used
is ML methods in the literature.

A new output result variable, "total_cost", is added to allow saving the OPF
solution total cost for each sample.

Variable names throughout the code and documentation have been updated to
account for the above changes.